VFEP-1330 Merge directDepositUltra with DirectDepositChangeUpdate.
