+++ src/applications/edu-benefits/pages/directDepositChangeUpdate.js
@@ -1,11 +1,22 @@
+import React from 'react';
 import get from 'platform/utilities/data/get';
 import set from 'platform/utilities/data/set';
+import bankAccountUI from 'platform/forms/definitions/bankAccount';
+import { isValidRoutingNumber } from 'platform/forms/validations';
 
-// import bankAccountUI from 'platform/forms/definitions/bankAccount';
-
-import React from 'react';
 import { bankAccountChangeLabelsUpdate } from '../utils/labels';
-import { isValidRoutingNumber } from '../utils/helpers';
+
+function validateRoutingNumber(
+  errors,
+  routingNumber,
+  formData,
+  schema,
+  errorMessages,
+) {
+  if (!isValidRoutingNumber(routingNumber)) {
+    errors.addError(errorMessages.pattern);
+  }
+}
 
 function isStartUpdate(form) {
   return get('bankAccountChangeUpdate', form) === 'startUpdate';
@@ -17,46 +28,6 @@ const gaBankInfoHelpText = () => {
     'help-text-label': 'What if I don’t have a bank account?',
   });
 };
-
-const directDepositDescriptionAssistance = (
-  <div className="vads-u-margin-top--2 vads-u-margin-bottom--2">
-    <p>
-      Direct deposit information is not mandatory at this time. However,
-      benefits cannot be awarded without this information per U.S. Treasury
-      regulation 31 C.F.R. § 208.3.
-    </p>
-    <p>For assistance call 1-888-442-4551 (1-888-GIBILL-1)</p>
-  </div>
-);
-
-const bankInfoHelpText = (
-  <>
-    {directDepositDescriptionAssistance}
-    <va-additional-info
-      trigger="What if I don’t have a bank account?"
-      onClick={gaBankInfoHelpText}
-    >
-      <span>
-        <p>
-          The{' '}
-          <a href="https://veteransbenefitsbanking.org/">
-            Veterans Benefits Banking Program (VBBP)
-          </a>{' '}
-          provides a list of Veteran-friendly banks and credit unions. They’ll
-          work with you to set up an account, or help you qualify for an
-          account, so you can use direct deposit. To get started, call one of
-          the participating banks or credit unions listed on the VBBP website.
-          Be sure to mention the Veterans Benefits Banking Program.
-        </p>
-        <p>
-          Note: Federal regulation, found in 31 C.F.R. § 208.3 provides that,
-          subject to section 208.4, “all Federal payments made by an agency
-          shall be made by electronic funds transfer” (EFT).
-        </p>
-      </span>
-    </va-additional-info>
-  </>
-);
 
 const directDepositDescription = (
   <div className="vads-u-margin-top--2 vads-u-margin-bottom--2">
@@ -70,6 +41,7 @@ const directDepositDescription = (
 const isFieldRequired = formData => {
   return formData.bankAccountChangeUpdate === 'startUpdate';
 };
+
 const directDepDescription = (
   <div>
     <p>
@@ -107,8 +79,10 @@ const directDepDescription = (
     </p>
   </div>
 );
-export default function createDirectDepositChangePage(schema) {
+
+export default function createDirectDepositPage(schema) {
   const { bankAccountChangeUpdate, bankAccount } = schema.definitions;
+
   return {
     title: 'Direct deposit',
     path: 'personal-information/direct-deposit',
@@ -124,18 +98,11 @@ export default function createDirectDepositChangePage(schema) {
           labels: bankAccountChangeLabelsUpdate,
         },
       },
-      'view:directDepositImageAndText': {
-        'ui:description': directDepositDescription,
-        'ui:options': {
-          hideIf: formData =>
-            formData.bankAccountChangeUpdate !== 'startUpdate',
-        },
-      },
       bankAccount: {
         'ui:order': ['accountType', 'routingNumber', 'accountNumber'],
+        'ui:description': directDepositDescription,
         'ui:options': {
           hideIf: formData => !isStartUpdate(formData),
-          expandUnder: 'view:directDepositImageAndText',
           updateSchema: (formData, _schema) =>
             set(
               'required',
@@ -156,24 +123,19 @@ export default function createDirectDepositChangePage(schema) {
           },
         },
         accountNumber: {
-          'ui:title': 'Bank account number',
+          ...bankAccountUI.accountNumber,
           'ui:errorMessages': {
-            required: 'Please enter your bank’s 9-digit routing number',
+            pattern: 'Please enter a valid account number',
+            required: 'Please enter a bank account number',
           },
         },
         routingNumber: {
-          'ui:title': 'Bank routing number',
-          'ui:validations': [isValidRoutingNumber],
+          ...bankAccountUI.routingNumber,
+          'ui:validations': [validateRoutingNumber],
           'ui:errorMessages': {
             pattern: 'Please enter your bank’s 9-digit routing number',
+            required: 'Please enter a routing number',
           },
-        },
-      },
-      'view:noneWarning': {
-        'ui:description': bankInfoHelpText,
-        'ui:options': {
-          hideIf: formData => formData.bankAccountChangeUpdate !== 'none',
-          expandUnder: 'bankAccountChangeUpdate',
         },
       },
     },
